requirements.txt
fastapi
uvicorn
torch
transformers
onnxruntime
numpy
kafka-python
pymongo
streamlit
pandas
altair